compiles on linux again

+++ src/DmScanLib.cpp
@@ -171,7 +171,7 @@ int DmScanLib::scanFlatbed(unsigned dpi, int brightness, int contrast,
 
 int DmScanLib::scanAndDecode(unsigned dpi, int brightness, int contrast,
 		BoundingBox<double> & region, const DecodeOptions & decodeOptions,
-		std::vector<std::unique_ptr<WellRectangle<unsigned>  > > & wellRects) {
+		std::vector<std::unique_ptr<WellRectangle<double>  > > & wellRects) {
 	VLOG(2) << "decodePlate: dpi/" << dpi << " brightness/" << brightness
 			<< " contrast/" << contrast
 			<< " " << region << " " << decodeOptions;
@@ -201,7 +201,7 @@ int DmScanLib::scanAndDecode(unsigned dpi, int brightness, int contrast,
 
 int DmScanLib::decodeImageWells(const char * filename,
 		const DecodeOptions & decodeOptions,
-		std::vector<std::unique_ptr<WellRectangle<unsigned>  > > & wellRects) {
+		std::vector<std::unique_ptr<WellRectangle<double>  > > & wellRects) {
 
 	VLOG(2) << "decodeImage: filename/" << filename
 			<< " numWellRects/" << wellRects.size()
@@ -219,7 +219,7 @@ int DmScanLib::decodeImageWells(const char * filename,
 int DmScanLib::decodeCommon(const Dib & image,
 		const DecodeOptions & decodeOptions,
 		const std::string &decodedDibFilename,
-		std::vector<std::unique_ptr<WellRectangle<unsigned>  > > & wellRects) {
+		std::vector<std::unique_ptr<WellRectangle<double>  > > & wellRects) {
 
 	decoder = std::unique_ptr<Decoder>(new Decoder(image, decodeOptions, wellRects));
 	int result = decoder->decodeWellRects();
