+++ src/decoder/Decoder.h
@@ -47,7 +47,7 @@ class DmtxDecodeHelper;
 class Decoder {
 public:
     Decoder(const Dib & image, const DecodeOptions & decodeOptions,
-    		std::vector<std::unique_ptr<WellRectangle<unsigned>  > > & wellRects);
+    		std::vector<std::unique_ptr<WellRectangle<double>  > > & wellRects);
     virtual ~Decoder();
     int decodeWellRects();
     void decodeWellRect(const Dib & wellRectImage, WellDecoder & wellDecoder) const;
@@ -79,7 +79,7 @@ private:
     const Dib & image;
     std::unique_ptr<Dib> workingImage;
     const DecodeOptions & decodeOptions;
-    const std::vector<std::unique_ptr<WellRectangle<unsigned> > > & wellRects;
+    const std::vector<std::unique_ptr<WellRectangle<double> > > & wellRects;
     std::vector<std::unique_ptr<WellDecoder> > wellDecoders;
     bool decodeSuccessful;
     std::map<std::string, const WellDecoder *> decodedWells;
